builtins: kill: Do not set signal on current shell when pid is empty
    % zsh; echo "$?"
    % trap 'exit 5' TERM
    % kill ''
    5

This behaviour seems more likely to be the result of bugs in programs
(eg. `kill -9 "$unsetvar") rather than be desirable behaviour to me. It
seems unintentional judging by the code and documentation, since it
comes about as a result of `isanum` returns true for empty strings
(since an empty string technically only consists of digits and
minuses...) and `atoi`, when passed a pointer to an invalid number,
returns 0.

With this patch:

    % trap 'exit 5' TERM
    % kill ''
    kill: empty pid provided